Move setting to room profile
