+++ vector/src/main/java/im/vector/app/features/roomprofile/settings/RoomSettingsAction.kt
@@ -28,7 +28,6 @@ sealed class RoomSettingsAction : VectorViewModelAction {
     data class SetRoomHistoryVisibility(val visibility: RoomHistoryVisibility) : RoomSettingsAction()
     data class SetRoomJoinRule(val roomJoinRule: RoomJoinRules) : RoomSettingsAction()
     data class SetRoomGuestAccess(val guestAccess: GuestAccess) : RoomSettingsAction()
-    data class SetEncryptToVerifiedDeviceOnly(val enabled: Boolean) : RoomSettingsAction()
 
     object Save : RoomSettingsAction()
     object Cancel : RoomSettingsAction()
