+++ vector/src/main/java/im/vector/app/features/roomprofile/settings/RoomSettingsFragment.kt
@@ -43,7 +43,6 @@ import im.vector.app.core.utils.toast
 import im.vector.app.databinding.FragmentRoomSettingGenericBinding
 import im.vector.app.features.analytics.plan.MobileScreen
 import im.vector.app.features.home.AvatarRenderer
-import im.vector.app.features.navigation.SettingsActivityPayload
 import im.vector.app.features.roomprofile.RoomProfileArgs
 import im.vector.app.features.roomprofile.RoomProfileSharedActionViewModel
 import im.vector.app.features.roomprofile.settings.historyvisibility.RoomHistoryVisibilityBottomSheet
@@ -198,14 +197,6 @@ class RoomSettingsFragment :
         val currentGuestAccess = state.newRoomJoinRules.newGuestAccess ?: state.currentGuestAccess
         val toggled = if (currentGuestAccess == GuestAccess.Forbidden) GuestAccess.CanJoin else GuestAccess.Forbidden
         viewModel.handle(RoomSettingsAction.SetRoomGuestAccess(toggled))
-    }
-
-    override fun setEncryptedToVerifiedDevicesOnly(enabled: Boolean) {
-        viewModel.handle(RoomSettingsAction.SetEncryptToVerifiedDeviceOnly(enabled))
-    }
-
-    override fun openGlobalBlockSettings() {
-        navigator.openSettings(requireContext(), SettingsActivityPayload.SecurityPrivacy)
     }
 
     override fun onImageReady(uri: Uri?) {
